+++ lib/sg_fargate_rails/config.rb
@@ -1,9 +1,9 @@
 module SgFargateRails
   class Config
-    attr_accessor :proxy_ip_address, :paths_to_allow_access_only_from_proxy
+    attr_accessor :proxy_ip_addresses, :paths_to_allow_access_only_from_proxy
 
     def initialize
-      self.proxy_ip_address = ENV['SG_PROXY_IP_ADDRESS']
+      self.proxy_ip_addresses = ENV['SG_PROXY_IP_ADDRESSES']
     end
   end
 end